SNOW-2183023: refactored
