+++ src/snowflake/connector/network.py
@@ -1,16 +1,12 @@
 #!/usr/bin/env python
 from __future__ import annotations
 
-import collections
-import contextlib
 import gzip
-import itertools
 import json
 import logging
 import re
 import time
 import uuid
-from collections import OrderedDict
 from threading import Lock
 from typing import TYPE_CHECKING, Any, Callable
 
@@ -18,10 +14,6 @@ import OpenSSL.SSL
 
 from snowflake.connector.secret_detector import SecretDetector
 from snowflake.connector.vendored.requests.models import PreparedRequest
-from snowflake.connector.vendored.urllib3.connectionpool import (
-    HTTPConnectionPool,
-    HTTPSConnectionPool,
-)
 
 from . import ssl_wrap_socket
 from .compat import (
@@ -84,6 +76,7 @@ from .errors import (
     ServiceUnavailableError,
     TooManyRequests,
 )
+from .session_manager import SessionManager, SessionPool
 from .sqlstate import (
     SQLSTATE_CONNECTION_NOT_EXISTS,
     SQLSTATE_CONNECTION_REJECTED,
@@ -96,19 +89,16 @@ from .time_util import (
 )
 from .tool.probe_connection import probe_connection
 from .vendored import requests
-from .vendored.requests import Response, Session
+from .vendored.requests import Response
 from .vendored.requests.adapters import HTTPAdapter
 from .vendored.requests.auth import AuthBase
 from .vendored.requests.exceptions import (
     ConnectionError,
     ConnectTimeout,
-    InvalidProxyURL,
     ReadTimeout,
     SSLError,
 )
-from .vendored.requests.utils import prepend_scheme_if_needed, select_proxy
 from .vendored.urllib3.exceptions import ProtocolError
-from .vendored.urllib3.poolmanager import ProxyManager
 from .vendored.urllib3.util.url import parse_url
 
 if TYPE_CHECKING:
@@ -248,42 +238,6 @@ def is_login_request(url: str) -> bool:
     return "login-request" in parse_url(url).path
 
 
-class ProxySupportAdapter(HTTPAdapter):
-    """This Adapter creates proper headers for Proxy CONNECT messages."""
-
-    def get_connection(
-        self, url: str, proxies: OrderedDict | None = None
-    ) -> HTTPConnectionPool | HTTPSConnectionPool:
-        proxy = select_proxy(url, proxies)
-        parsed_url = urlparse(url)
-
-        if proxy:
-            proxy = prepend_scheme_if_needed(proxy, "http")
-            proxy_url = parse_url(proxy)
-            if not proxy_url.host:
-                raise InvalidProxyURL(
-                    "Please check proxy URL. It is malformed"
-                    " and could be missing the host."
-                )
-            proxy_manager = self.proxy_manager_for(proxy)
-
-            if isinstance(proxy_manager, ProxyManager):
-                # Add Host to proxy header SNOW-232777
-                proxy_manager.proxy_headers["Host"] = parsed_url.hostname
-            else:
-                logger.debug(
-                    f"Unable to set 'Host' to proxy manager of type {type(proxy_manager)} as"
-                    f" it does not have attribute 'proxy_headers'."
-                )
-            conn = proxy_manager.connection_from_url(url)
-        else:
-            # Only scheme should be lower case
-            url = parsed_url.geturl()
-            conn = self.poolmanager.connection_from_url(url)
-
-        return conn
-
-
 class RetryRequest(Exception):
     """Signal to retry request."""
 
@@ -332,101 +286,6 @@ class PATWithExternalSessionAuth(AuthBase):
         if self.external_session_id != NO_EXTERNAL_SESSION_ID:
             r.headers[HEADER_EXTERNAL_SESSION_KEY] = self.external_session_id
         return r
-
-
-class SessionManager:
-    def __init__(
-        self,
-        use_pooling: bool = True,
-        adapter_factory: (
-            Callable[..., HTTPAdapter] | None
-        ) = lambda *args, **kwargs: None,
-    ):
-        self._use_pooling = use_pooling
-        self._adapter_factory = adapter_factory or ProxySupportAdapter
-        self._sessions_map: dict[str | None, SessionPool] = collections.defaultdict(
-            lambda: SessionPool(self)
-        )
-
-    @property
-    def sessions_map(self) -> dict[str, SessionPool]:
-        return self._sessions_map
-
-    def _mount_adapter(self, session: requests.Session) -> None:
-        adapter = self._adapter_factory(max_retries=REQUESTS_RETRY)
-        if adapter is not None:
-            session.mount("http://", adapter)
-            session.mount("https://", adapter)
-
-    def make_session(self) -> Session:
-        s = requests.Session()
-        self._mount_adapter(s)
-        s._reuse_count = itertools.count()
-        return s
-
-    @contextlib.contextmanager
-    def use_session(self, url: str | None = None):
-        if not self._use_pooling:
-            session = self.make_session()
-            try:
-                yield session
-            finally:
-                session.close()
-        else:
-            hostname = urlparse(url).hostname if url else None
-            pool = self._sessions_map[hostname]
-            session = pool.get_session()
-            try:
-                yield session
-            finally:
-                pool.return_session(session)
-
-    def close(self):
-        for pool in self._sessions_map.values():
-            pool.close()
-
-
-class SessionPool:
-    def __init__(self, manager: SessionManager) -> None:
-        # A stack of the idle sessions
-        self._idle_sessions = []
-        self._active_sessions = set()
-        self._manager = manager
-
-    def get_session(self) -> Session:
-        """Returns a session from the session pool or creates a new one."""
-        try:
-            session = self._idle_sessions.pop()
-        except IndexError:
-            session = self._manager.make_session()
-        self._active_sessions.add(session)
-        return session
-
-    def return_session(self, session: Session) -> None:
-        """Places an active session back into the idle session stack."""
-        try:
-            self._active_sessions.remove(session)
-        except KeyError:
-            logger.debug("session doesn't exist in the active session pool. Ignored...")
-        self._idle_sessions.append(session)
-
-    def __str__(self) -> str:
-        total_sessions = len(self._active_sessions) + len(self._idle_sessions)
-        return (
-            f"SessionPool {len(self._active_sessions)}/{total_sessions} active sessions"
-        )
-
-    def close(self) -> None:
-        """Closes all active and idle sessions in this session pool."""
-        if self._active_sessions:
-            logger.debug(f"Closing {len(self._active_sessions)} active sessions")
-        for session in itertools.chain(self._active_sessions, self._idle_sessions):
-            try:
-                session.close()
-            except Exception as e:
-                logger.info(f"Session cleanup failed - failed to close session: {e}")
-        self._active_sessions.clear()
-        self._idle_sessions.clear()
 
 
 # Customizable JSONEncoder to support additional types.
